- added Halflife 1 Model Loader - added Halflife 1 Texture Loader - slightly changed usage of getMesh ( doesn't effect existing implementation)
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@3335 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CDepthBuffer.cpp
@@ -122,7 +122,7 @@ void CStencilBuffer::setSize(const core::dimension2d<u32>& size)
 	if (Buffer)
 		delete [] Buffer;
 
-	Pitch = size.Width * sizeof ( u8 );
+	Pitch = size.Width * sizeof ( u32 );
 	TotalSize = Pitch * size.Height;
 	Buffer = new u8[TotalSize];
 	clear ();
